+++ {{cookiecutter.repo_name}}/app/src/sagas/index.js
@@ -3,10 +3,12 @@ import { getLocation, replace } from 'connected-react-router';
 import { fork, put, select } from 'redux-saga/effects';
 
 import routes from 'configuration/routes';
+import refreshTokenManager from 'sagas/auth/refreshTokenSaga';
 
 
 export default function* rootSaga(hot = false) {
-    yield fork(ViewManager, routes);
+    yield fork(refreshTokenManager);
+    yield fork(ViewManager, routes, { allowLogger: true });
 
     if (hot) {
         const location = yield select(getLocation);